Correccion error tipo de dato en puntajes

+++ mecano-4a-ap-gateway/src/typeDefs/puntajes_type_defs.js
@@ -7,7 +7,7 @@ const puntajesTypeDefs = gql(`
         leccionId   : String!
         precision   : Float!
         cpm_e       : Int!
-        segundos    : Int!
+        segundos    : Float!
         fecha       : String
     }
 
@@ -17,7 +17,7 @@ const puntajesTypeDefs = gql(`
         leccionId   : String
         precision   : Float
         cpm_e       : Int
-        segundos    : Int
+        segundos    : Float
         fecha       : String
     }
